AB#10938

+++ src/SME.SGP.Dominio/Entidades/RecuperacaoParalela.cs
@@ -2,8 +2,8 @@
 {
     public class RecuperacaoParalela : EntidadeBase
     {
-        public int Aluno_id { get; set; }
+        public long Aluno_id { get; set; }
         public bool Excluido { get; set; }
-        public int TurmaId { get; set; }
+        public long TurmaId { get; set; }
     }
 }
